Don't allow an implicit conversion from NSNumber to one of its bridged value types.
Eliminate the implicit conversion from NSNumber to one of its bridged
value types (Int, Float, Double, Bool, etc.). One can use "as" to
perform the conversion instead. This eliminates a class of
accepts-dubious that involve lossy conversions from NSNumber to Int.

Our eventual arc is to eliminate all of these conversions. This is a
small, high-value step along that path rdar://problem/18269449.

Swift SVN r21879

+++ test/Interpreter/SDK/CoreGraphics_CGFloat.swift
@@ -127,7 +127,7 @@ CGFloatTestSuite.test("bridging") {
     expectEqual("4.125", "\(nsnum)")
 
     // NSNumber -> CGFloat
-    let bridgedBack: CGFloat = nsnum
+    let bridgedBack: CGFloat = nsnum as CGFloat
     expectEqual(flt, bridgedBack)
   }
